Move Home class to pages directory
