+++ Framework/Pages/Home.cs
@@ -2,7 +2,7 @@
 using OpenQA.Selenium.Support.UI;
 using System;
 
-namespace Framework
+namespace Framework.Pages
 {
     public class Home
     {
